+++ voyage_list.php
@@ -180,7 +180,7 @@ $arrayfields = dol_sort_array($arrayfields, 'position');
 
 // There is several ways to check permission.
 // Set $enablepermissioncheck to 1 to enable a minimum low level of checks
-$enablepermissioncheck = 0;
+$enablepermissioncheck = 1;
 if ($enablepermissioncheck) {
 	$permissiontoread = $user->hasRight('clienjoyholidays', 'voyage', 'read');
 	$permissiontoadd = $user->hasRight('clienjoyholidays', 'voyage', 'write');