+++ ChangeLog.md
@@ -2,5 +2,6 @@
 
 ## 1.0
 
+- NEW - Added permission to read/delete/write "voyage" - *17/08/2023* - 1.3
 - NEW - Add label condition (5 caract min) - *17/08/23023* - 1.2 
 - NEW - Add "voyage" object + transport mode dictionary + picto - *16/08/2023* - 1.1